moved assembly helpers to separate file
